Cast double to int

+++ app/src/main/java/com/k3/rtlion/ScannerPageFrag.java
@@ -182,11 +182,11 @@ public class ScannerPageFrag {
                 throw new Exception(context.getString(R.string.invalid_dbvals));
             int dbSum = 0;
             for (int i = 0; i < dbRes.size(); i++) {
-                dbSum += Integer.parseInt(dbRes.get(i));
+                dbSum += (int) Double.parseDouble(dbRes.get(i));
             }
             int dbAvg = dbSum / dbRes.size();
             for (int i = 0; i < freqRes.size(); i++) {
-                if (Math.abs(Integer.parseInt(dbRes.get(i))) <= Math.abs(dbAvg / 2))
+                if (Math.abs((int) Double.parseDouble(dbRes.get(i))) <= Math.abs(dbAvg / 2))
                     freqRes.remove(freqRes.get(i));
             }
             arrayAdapterRes.notifyDataSetChanged();
